requirements.txt
numpy==1.26.4
PyYAML==6.0.2
torch==2.3.1
